Cleaned up some Button styling
Cleaned up a little bit of the Button styling. Still more work that
could be done here, since it seems that the bbUI.js style isn't quite in
line with what Cascades actually looks like. We should be able to
achieve visual parity with Cascades, but it will take some fine tuning.