Preliminary attempt at a BB10 Slider
First attempt at a BB10 Slider in pure QML. The framework for the
control's look and feel are there, but there's still a lot to implement.
It seems that we'll need to use DropShadows and RadialGradients to
achieve the true look we want, which will require the use of shaders.